# Conflicts:
#	pom.xml
#	src/main/java/com/blueteam/bluequiz/config/MvcConfig.java
#	src/main/java/com/blueteam/bluequiz/controller/Api.java
#	src/main/java/com/blueteam/bluequiz/controller/UserController.java
#	src/main/java/com/blueteam/bluequiz/entities/UserAnswersContainer.java
#	src/main/java/com/blueteam/bluequiz/security/WebSecurityConfig.java
#	src/main/java/com/blueteam/bluequiz/service/QuizManagerService.java
#	src/main/java/com/blueteam/bluequiz/service/QuizResultService.java
#	src/main/resources/templates/login.html

+++ src/main/java/com/blueteam/bluequiz/service/QuizManagerService.java
@@ -2,12 +2,16 @@ package com.blueteam.bluequiz.service;
 
 import com.blueteam.bluequiz.entities.Quiz;
 import com.blueteam.bluequiz.persistence.QuizRepository;
+import lombok.extern.log4j.Log4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Log4j
 @Service
 public class QuizManagerService {
 
@@ -27,7 +31,8 @@ public class QuizManagerService {
     }
 
     public void insert(Quiz quiz) {
-        Quiz newQuiz = Quiz.builder()
+        if (hasRoleAdmin()) {
+             Quiz newQuiz = Quiz.builder()
                 .quizTitle(quiz.getQuizTitle())
                 .questions(quiz.getQuestions())
                 .build();
@@ -35,7 +40,8 @@ public class QuizManagerService {
     }
 
     public void update(String id, Quiz quiz) {
-        if (quizRepository.existsById(id)) {
+        if (hasRoleAdmin()) {
+                if (quizRepository.existsById(id)) {
             quizRepository.save(Quiz.builder()
                     ._id(id)
                     .quizTitle(quiz.getQuizTitle())
@@ -44,6 +50,17 @@ public class QuizManagerService {
     }
 
     public void deleteById(String id) {
-        quizRepository.deleteById(id);
+        if (hasRoleAdmin()) {
+               quizRepository.deleteById(id);
+    }
+}
+
+    private boolean hasRoleAdmin() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!authentication.getAuthorities().toArray()[0].toString().equals("ADMIN")) {
+            log.debug("User " + authentication.getName() + " try to call this method without grant ADMIN");
+            return false;
+        }
+        return true;
     }
 }
